requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.7
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-oauth==1.0.1
googleapis-common-protos==1.62.0
httplib2==0.22.0
icecream==2.1.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
oauthlib==3.2.2
platformdirs==3.11.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
yarl==1.9.4
